LWO  - added workaround for LWOB's with ill-formed SURF chunks  - layer hierarchy is now correct  - fixed & simplified transparency handling
FindDegenerates 
 - fixed seldom crashes with RemoveDegenerates=1. 


git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@369 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/LWOLoader.h
@@ -314,7 +314,7 @@ private:
 	 *  @param parent Index of the node
 	 *  @param apcNodes Flat list of nodes - used nodes are set to NULL.
 	*/
-	void AddChildren(aiNode* node, uintptr_t parent, 
+	void AddChildren(aiNode* node, uint16_t parent, 
 		std::vector<aiNode*>& apcNodes);
 
 	// -------------------------------------------------------------------
